requirements.txt
annotated-types==0.7.0
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
PyYAML==6.0.2
typing-inspection==0.4.1
typing_extensions==4.15.0
